Basic changes for creating a Matcher client separating logic from handler
