+++ WorkloadManagementSystem/Client/JobState/JobState.py
@@ -1,3 +1,7 @@
+""" What's this...?
+"""
+
+__RCSID__ = "$Id"
 
 import types
 from DIRAC import gLogger, S_OK, S_ERROR
@@ -6,8 +10,7 @@ from DIRAC.WorkloadManagementSystem.Client.JobState.JobManifest import JobManife
 from DIRAC.Core.DISET.RPCClient import RPCClient
 from DIRAC.WorkloadManagementSystem.Service.JobPolicy import RIGHT_GET_INFO, RIGHT_RESCHEDULE
 from DIRAC.WorkloadManagementSystem.Service.JobPolicy import RIGHT_RESET, RIGHT_CHANGE_STATUS
-
-__RCSID__ = "$Id"
+from DIRAC.WorkloadManagementSystem.DB.TaskQueueDB import singleValueDefFields, multiValueDefFields
 
 class JobState( object ):
 
@@ -514,14 +517,14 @@ class JobState( object ):
     reqCfg = result[ 'Value' ]
 
     jobReqDict = {}
-    for name in JobState.__db.tq.getSingleValueTQDefFields():
+    for name in singleValueDefFields:
       if name in reqCfg:
         if name == 'CPUTime':
           jobReqDict[ name ] = int( reqCfg[ name ] )
         else:
           jobReqDict[ name ] = reqCfg[ name ]
 
-    for name in JobState.__db.tq.getMultiValueTQDefFields():
+    for name in multiValueDefFields:
       if name in reqCfg:
         jobReqDict[ name ] = reqCfg.getOption( name, [] )
